fix(input): remove xinput pointer definition on build without XInput
The XInpupt dependency is optional and in case it's not available some code
parts are deactivated with precompiler instructions.

In this case we also need to deactivate the smart pointer to the xinput class
too as otherwise we get a build error on the smart pointer not having the
definition available.

Do this and restructure the function definition for its creation as we can not
rely on the symbol being available. Just define an else case with a separate
function definition.