fix: configure xdg-toplevels with Plasma shell surface too
A Plasma shell surface which is also an xdg-toplevel should still use the
synchronization sequence of xdg-toplevels for its size. In this case we are
required to send a configure event before adjusting the geometry.

This fixes the assert firing on the maximization mode not yet being
synchronized when a window of the app Telegram is maximized.

As we now might configure the geometry of a Plasma shell surface we must ensure
that in such a case we at least update the position of the pending and current
frame geometry. This leads to unsynchronized positional changes of the window
but this is to be expected as the Plasma shell protocol is ignorant of such.