fix: discard quads on position changes Plasma shell surface popups
A Plasma shell surface might be created for a wl_surface which also has the
xdg-popup role. In that case the window is set as annexed to its xdg-popup
parent while we might still do explicit repositioning through the Plasma shell
surface mechanism.

On such a reposition we have to discard the quads of the parent, as we paint
the annexed child surface always as quads of the parent.

This fixes the moving of taskbar tooltips.