fix: check on surface when setting Xwayland window surface
The out-of-sync communication with Xwayland can lead to a wl_surface being
created only after the Xwayland window has been via the X11 connection. In this
case we must associate the window and the surface.

But we need to be careful in what window we choose for this. We had an issue
with some other window having its surface reset instead of a Xwayland window in
case the surface ids collide. We fix this by checking that a surface has not
yet been set on the Toplevel. For Wayland native windows the surface is an
invariant. It is always set and we won't choose it for setting its surface.