build: clean up .gitignore entries
Files specific to user IDEs or workflows should not be listed in the .gitignore
file of the repo. Instead the user should add them to a global .gitignore file.

We do a cleanup of the .gitignore file, removing user-specific files and
instead adding files which are common to building the project and our tooling.

A comment informs contributors about the posssibilitiy of creating a global
.gitignore file for their personal editor files instead of using the repo one.