fix: always check on input device when printing out information about it
Various input events might come without a device specified. This is true for
at least button events and also motion events (pushback when approaching
screenedges).

Instead of checking the argument in each call, just provide the device itself
to the function, now as a template over the devices, and check inside the
function.