build: remove Doxygen main page files
These files can be useful when automatically generating Doxygen documentation,
but for our main binary it does not make sense to generate such. For the
effects library documentation could be generated in a meaningful way, but the
usefulness is limited also in this case.

In any case the additional information provided by the main pages is minimal
and there is currently no process established by us to generate and publish
the documentation, but which would probably not be very useful after all. So it
makes overall sense to remove the files.