feat: record render time
In order to delay our buffer presentation shortly before vblank and keeping it
from being too late under high rendering load, record the render times of the
last 100 frames as we do with our internal paint time already.

We do this with OpenGl timer queries. On each run the previous ones can be read
out. This way we don't need to block or wait on it in a separate thread.