fix: disable rendering on suspended outputs
We should stop trying to render on an output which is disabled via DPMS. We do
this by switching off timers on the output. After reenabling we have to provide
a first buffer for wlroots to enable it again. For that we issue a full repaint
of the output and then enable it before the end of its paint pipeline.