refactor: adapt to Wraplang setting the selection source
Instead of setting the data device we now only set the source in Wrapland for
selections. This simplifies the code but we need to adjust our Xwayland
integration code quite a bit. It is important to adapt not only the interface
but also replace some connections to get updated source information correctly.