fix: get refresh length in nanoseconds
The refresh rate must be returned in nanoseconds for the delay algorithm to
function. Note that while this fixes our calculation it worsens the user
experience for the moment as we do not yet factor in the render time after our
internal paint process finished. This will be fixed in a follow-up patch.