refactor(xwl): transform wl_source member functions to free functions
This way we can get rid of the template specialization. The functions are moved
into the newly created selection_wl.h header. We change up their signature to
provide the necessary arguments to the free functions.

In the start transfer function we return the file descriptor of write side of
the pipe instead of a boolean so the caller can write into it and -1 in case of
error.