refactor(xwl): transform x11_source member functions to free functions
This way we can get rid of the template specialization. The functions are moved
into the newly created selection_x11.h header. We change up their signature to
provide the necessary arguments to the free functions.

The sources.cpp file can be removed. We compile the header instead for the
QObjects defined in there.