refactor(input): virtualize filter setup
We move filter creation to the child classes of redirect. This way we may pick
each filter in the respective child class acording to what's needed for the
windowing system.

Filters that are only needed in the Wayland child class we directly compile
into the Wayland libary only. Filters needed in both are exported from the
standard "kwin" library.