refactor: set Xwayland surfaces from Wayland server
Add a signal to the workspace that Xwayland windows are supposed to emit when
their surface id got changed via an X11 message. This signal emittion is picked
up by the Wayland server class and the surface is set.

This way we can reduce the dependency of the Toplevel class on Wayland server.