docs: recommend pipeline timeout default adjustment
The build job in general takes longer than one hour with the shared GitLab
runners.

For new contributors, who just forked the repo, the pipeline timeout is set to
one hour by default though. Recommend in the Contributing document to manually
adjust this default so that pipelines are not failing early.