refactor: compile the server class into the Wayland library
In a major change we restrict the compilation of the WaylandServer to the
Wayland library.

To be able to do this we must simultaneously restrict the compilation of the
AbstractWaylandOutput class to the Wayland library since it accesses
WaylandServer. For that we transform Platform functions relating to the DPMS
handling of our Wayland outputs to free function templates in the base
namespace and adapt the DPMS input filter accordingly.