refactor(space): setup Xwayland Plasma management in Wayland space
We don't want to reference Wayland functionality in the win::x11 namespace.
Instead handle new Xwayland windows from the Wayland space. We do this when
the Xwayland windows have become ready to paint, which is similar to when
the setup function was before called directly in the win::x11 namespace.