refactor(input): store base platform in Wayland platform
Store a reference to the base platform of our Wayland session in the Wayland
input platform. This will help us with later changes so to be able to talk to
the base from the input platform.

For the moment we specify the wlroots backend type explicitly but hope to be
able to make this more generic through templating in the future.