feat: remove support for KDE Activities
Activities have usability issues and technical flaws. How they are integrated
into the compositor is problematic. Activities also are used only by very few
users. Activities are complicated to setup and maintain while virtual desktops
are simpler to handle and offer a large subset of functionality that is more
than enough for the majority of people.

The bottom line is that KDE Activities simply don't justify the additional
complexity their implementation comes with.

Support for Activities is therefore removed with this commit. We purge most
references to them, but for scripts and effects keep around some public
interfaces resulting in noops to ensure backwards compatibility.