refactor: manage all Wayland windows in space
Move all management of announced Wayland windows from the WaylandServer class
to the win::wayland::space class as this is where windows are supposed to be
managed.

This removes the back-and-forth between the workspace and the server class on
window creation. We have a single functin to handle a new window in the space
class and call it either directly if the window is ready to be painted or when
it is first shown.