refactor(space): store Wayland windows directly in windows container
Remove the additional container to store Wayland windows in the Wayland space
and instead directly store them only in the m_windows container of the base
class.

This simplifies our internal data handling. For now we sometimes require
additional casts though to act on the windows. Our plan is to later on use
std::variant consistently for that.