refactor: move internal flags type to the KWin namespace
The flags type and our ENUM_FLAGS macro were introduced in the win namespace.
To make use of it we move it to a new toplevel directory "utils", that in the
future will be popuplated with more common utility types being useful in all
areas of the compositor.

It is observed that our flags mechanism clashes with the QFlags operators. To
fix that we directly convert some of our previous QFlags and where this is not
possible or kept for now for other reasons, we explicitly create a QFlags type
with initializer lists, which then compiles.