feat(input): introduce XKB keymaps per keyboard
We add a new class xkb_keyboard, which holds all keyboard-individual XKB types,
primarily the state, keymap and layout. An xkb_keyboard is created for every
keyboard device.

For compatibility and to handle the case when no keyboard is connected a
default keyboard is created in the xkb class, which is only a holder for that
default keyboard and the context anymore.