test: check layouts on multiple keyboards
We check for creation of additional keyboard devices and that signals are
correctly transmitted on the new D-Bus v2 interface.

To test additional keyboards we allow tests to access the wlroots backend in
the base class and adapt the keyboard press/release test helper functions to
allow tests to test keys on a different keyboard than the one created by
default.