fix(input): update keymap on every key and modifier event
When multiple keyboards with different keymaps are connected on the same seat
we must ensure that clients on every keystroke always have an updated keymap
for the next key.

On the other side we don't need to send a keymap if one of the keyboards simply
receives an updated one. We wait until it is actually sending the first key or
modifier.

A new integration test is added to control our internal keymap processing.