fix(space): ignore OSD position hints and configure events
There is a regression due to a recent change in Plasma Workspace. [1] The OSD
for e.g. the volume feedback was not centralized anymore in an X11 session. The
reason for that is that Plasma Workspace with [1] is now sending a geometry
hint for the OSD and afterwards configuring it to the (0,0) position. We deal
with that by checking for the window being an OSD and in that case always
placing the window and aborting later configure events.

[1] https://invent.kde.org/plasma/plasma-workspace/-/commit/e4ea7286