style: clang-format scene and effect files
We clang-format files related to our scene in preparation of a later move to
the render directory where we already have clang-format enforced on.

We don't want to change up the drm_fourcc.h header, so we move it to a new
toplevel directory "extern", which is as with Wrapland meant for such external
files that are not adhering to our own coding standard and should not differ
from the original files.