refactor(scene): make paint mask enum class
Instead of having an enum in the scene class, define an enum class outside that
mirrors the enum type in the effects lib, but with type safety. Internally we
use that enum class exclusively and cast it to int only when interacting with
the effects API in order to keep it stable.