refactor(scene): remove overlay window getters
The virtual overlay window getter function structure bleeded an implementation
detail of the X11 compositing into the generic parts of our internal
abstraction interface.

As we now have a pointer to the overlay window in the x11::compositor class we
can correct this architectural failure and get rid of the getters.