feat(scene): merge the OpenGL 2 and 3 backends
This change merges the two OpenGL backends into one making the current
default of GLCore the overall default. It becomes the first context to
try to create. If it fails, it will automatically fall back to the
(previous) OpenGL 2 backend. The change is inspired by a recent commit
in KWin [1].

[1] https://invent.kde.org/plasma/kwin/-/commit/68589fe9