refactor(scene): flatten OpenGL scene hierarchy
The gl::scene2 child class was introduced nine years ago in 5a6d9400 when there
was still code around for OpenGl 1. Today we only need a single child class for
OpenGL and this likely won't change in the future.

To simplify the code merge the scene2 class into its parent class.