feat: update launch environment via job
Plasma since recently [1] requires us to update the launch environment via an
UpdateLaunchEnvironmentJob and publish a service name on DBus to indicate that
the job has finished.

We do this simply at the end of our startup. Also when we run without the
wrapper as this should just be ignored then. Our wrapper on the other side we
keep as it is as a simple C program. In fact it would make sense to downsize it
further as it is rather useless, and on the contrary makes debugging difficult.

[1] https://invent.kde.org/plasma/plasma-workspace/-/commit/7884f53d