feat: remove Wayland wrapper binary
The introduction of the wrapper binary was an attempt to make the Wayland
session more "resilient" against compositor crashes.

In a nutshell the wrapper binary is spawned on Plasma session start instead of
the normal kwin_wayland binary, and then the wrapper executes the normal one
including reexecuting it in case it returns with a none-zero exit code.

There are multiple issues with this idea. Most of all for the foreseeable
future this can't work satisfyingly due to clients not recovering and the user
being confronted with the session restarted but everything important gone.

Another problem is that with the wrapper the complexity increases for little
gain. Having a session crash with restart is still a subpar user experience.
We should rather focus on improving the robustness of the compositor itself.

For these reasons and more we remove the wrapper binary. Instead of the wrapper
binary we install a symbolic link to our normal binary so that the Plasma
startup routines can still execute KWinFT as part of the Wayland session.