refactor: remove functions for setting initial Wayland data
In the past we used multiple functions to set initial data for a nested Wayland
session, the initial output count, size and scale.

Setting the output count is with our wlroots backend now possible via the env
vars WLR_WL_OUTPUTS and WLR_X11_OUTPUTS.

For the other data there is currently no way to set it. The size does not make
sense anymore as the wlroots backend comes with resizable nested outputs. For
the scale we could hook up a profile via Disman.

In any case nothing of the data is currently hooked up to be used with our
wlroots backend, so we remove all infrastructure for setting it.