refactor: dynamically inherit base platform
The strategy with the backend template is scrapped for now. Instead the backend
dynamically inherits the base platform. For Wayland and X11 we create a second
level in between what minimizes the number of casts we have to do for outputs.

Long term it might still be better to work with templates. But for that we
first have to ensure the classes don't have to inherit QObject. That won't be
possible in the short term though as the platform will have to inherit it.