style: namespace Night Color to render
The Night Color functionality can be classified as a post-process method to
manipulate the rendering. As such it is better suited to be part of the render
namespace. We create a new namespace for it and adapt the class names to
snake-case while at it.

In another commit we'll move the files to the subdirectory.