style: move screen edges support to win namespace
The Edge and ScreenEdges classes are renamed and moved to the win namespace as
these are concepts related to window and space management. The classes are
renamed to adapt to our overall code styling and to come with a succinct
meaning in the new namespace. The Edge class is renamed to screen_edge and the
manager class ScreenEdges is renamed to screen_edger.