refactor: update X11 time only in Application class
Move the update X11 time function from Platform to Application. Long term we
might be more interested in moving all update functions to templates in the
base namespace but for now we keep them all together in the Application.

The wrapper function in the utils header is removed too. Instead we do call the
member function - now in the Application class - directly.