refactor: rework determining of scripted effect config
Currently the plugin needs to set a property in oder to be considered
configurable.

In kpackage we have the concept of optional and required files, based on the
existence of those optional files we can define different behavior. For example
in Plasma the applets can optionally have a config ui.

Because here we don not use The KPackage:Package class, it is simpler to check
the existence of the files on disk.