refactor: flatten egl backend and texture class hierarchy
Remove the intermediate egl_backend and egl_texture classes and inherit the
Gl backend and private texture directly from the backend classes. This
simplifies our internal rendering structure.

To make this work we introduce a new egl_dmabuf_data struct and generate this
in the now templated dmabuf factory such that the dmabuf instance only takes
this data instead of a pointer back to the backend. That has the addtional
advantage of less coupling.