feat: render only with OpenGL ES on Wayland
On Wayland we may use the OpenGL or OpenGL ES graphics API to render. But for a
compositor GLES should be more than sufficient. As in the future we want to
further expand usage of wlroots in our rendering, and the wlroots renderer
only supports GLES, this becomes necessary too.

The OpenGL ES integration test is removed, since this is now covered directly
by the OpenGL test.