feat: use wlroots renderer swap chain
We replace our own EGL code and buffer swapping with wlroots API calls. This
allows us to remove a lot of our own backend code. As the wlroots renderer does
requires a GPU to be present for its buffer sharing many integration tests will
fail on CI pipelines without one. This will be improved upon in the future by
bringing back the software rendering backend and by trying to lift the
limitation in wlroots as being issued in [1].

[1] https://gitlab.freedesktop.org/wlroots/wlroots/-/issues/2871