fix: create DRM leases for non-desktop outputs only
The sole usecase of DRM leases are VR headsets at the moment. Other outputs are
in general meant to not be leased out but represented as a wl_output for normal
desktop usage.

wlroots indicates outputs for VR headsets with the non_desktop flag. For such
wlr_outputs we add a new base output type non_desktop_output that simply wraps
it but doesn't do much more. In particular we don't create a render output.

As this is completely indepedent from the render module, we moved the DRM lease
functionality to the base module previously and now create only for non-desktop
outputs DRM lease connectors. This overall also simplifies our code.