fix(scene): update software cursor on paint
A cursor change can happen outside of the repaint loop without an EGL context
current. Update the cursor texture only on the next repaint when a context does
exist for sure. For that we set a dirty flagon the cursor change but keep the
old one set until the next time in the repaint loop we would paint it.