refactor: handle wlroots frame event
Besides wlroot's present event on the backend we should also listen for the
frame event of backends. This way we can avoid retriggering the pipeline when
the session has become inactive. In this case wlroots only sends the present
but not the frame event.

Besides that from a theoretic standpoint rendering only on the frame event is
more conformant with requirements and in the future with more integration of
wlroots we'll omit running into issues by handling the frame event correctly.