ci: disable tests requiring GL
The wlroots renderer is at the moment not capable to run GL without a GPU as it
requires to attain a DRM render FD for dmabuf buffer passing. [1]

As we still want to continue pushing forward the integration of the wlroots
renderer for now disable on CI all integration tests that require GL to run.
The tests can still be run locally and hopefully soon again on CI after the
wlroots renderer has been improved in this regard.

[1] https://gitlab.freedesktop.org/wlroots/wlroots/-/issues/2871