refactor(scene): adapt QPainter backend to current architecture
The QPainter backend was decommissioned with the wlroots backend work to
simplify that transition. As the backend code has now been stabilized a lot in
comparison to last year we can add the QPainter rendering mode back again.

For that we first adapt the still present QPainter scene and abstract backend
interface to teh current architecture. In another step we'll add the actual
implementation in the wlroots backend.