refactor: init the workspace X11 integration separately
Instead of calling the init function from within the base class ctor and if not
yet an X11 connection is established connect to the connection changed signal,
call the init function from the outside in a Wayland session when the X11
connection is guaranteed to exist, i.e. after Xwayland has been launched.

On X11 call the function from within the child class win::x11::space.