refactor: store reference to workspace in XCB event filter
We can create the filter at such a time that a workspace is guaranteed to
exist and we can store a reference to the space in the filter. But we need to
ensure the filter is deleted together before the workspace. In the future
composite the filter inside the workspace.