refactor: setup compositing in free function template
We add more type guarantees this way. It'll help us in the future to further
depend on simple types.

We add a simple constexpr boolean in our window classes to static-assert that
we don't accidentally call some function on the parent class when we don't
expect it. This boolean is supposed to be only temporary and will be removed
again later on or go away together with the Toplevel class itself.