refactor: let Toplevel own render window and render window the effect window
We turn the ownership of render windows around. The Toplevel owns the render
window (scene window) when compositing is active. The render window in return
owns the effect window for integration with the effect system.

This way we have a clear ownership hierarchy and can tighten our assumptions in
many cases when accessing the render-related properties of windows.

When a window gets destroyed the render window is being transferred to the
remnant window. That implies compositing is active, otherwise no remannt is
being created in the first place.