refactor(scene): reset shadow toplevel directly
Remove the shadow::setToplevel function. That can be done in the scene directly
and it is more coherent as we do something very specific to the single call.

In shadow::updateShadow we assert on the member variable being set. It's clear
that it should always be non-null at this point.