feat(effect): do not highlight windows based on mouse if user is currently typing
If the user is currently typing, do not highlight windows based on the mouse
position. If the user types a window title before the effect opening animation
is finished, he ends up in a situation where only one window is on the screen
but it is deselected (because the mouse is not on the window) or another window
is selected because of the mouse. With this change, the mouse input does not
overwrite the window highlighting based on the search entered **if the user is
currently typing.**