feat: add environment variable to log stderr to file
When the environment variable KWIN_LOG_PATH is set, try to redirect stderr to
the path specified by that variable.

This is a workaround for our logs missing in the journal for a Plasma session
when this session is being launched from a VT via
dbus-run-session startplasma-wayland.

Our D-Bus activated user service should write its logs to the journal, but
doesn't from the secondary D-Bus session. This  seems to be a D-Bus or systemd
problem, see also [1].

[1] https://github.com/systemd/systemd/issues/22242