revert: "test: remove QPainter scene tests"
With the QPainter backend being added again, add back the integration tests
that were testing QPainter specifically.

The tests are adapted to compile and pass again despite the large-scale changes
to our code base in the meantime. Two checks about the buffer image on cursor
changes are for now set to QEXPECT_FAIL. These failing checks must be looked in
later on.

This reverts commit 44b077bd51621f507bb899ed1f76e525a56cf72c.