refactor(debug): move console files deeper
Create a directory "console" in the debug directory and move the console files
in there. This way we can reuse the debug directory for other tooling in the
future.

We use this opportunity to also move the X11 and Wayland specific files into
additional subdirectories to make it more tidy. The X11 console class is
compiled only into the X11 library anymore.