refactor: create wrapper atom class without default arguments
Default arguments should be omitted. Instead require to provide a connection to
the Xcb::Atom wrapper class. This also removes access to a global variable and
makes the code more generic.

The Atoms ctor is now called with a connection argument as well so it can
forward the argument to the individual atoms.