feat: omit setting the KWIN_RUNNING atom on the root window
The atom is being set by us at least since f52b8e48 but there is no reference
to it in other parts of KDE. It is presumed that setting the atom has become
unnecessary and we can remove the code without problems.

In its place, if at all, nowadays the XDG_CURRENT_DESKTOP environment variable
is used to identify the desktop environment.