feat: remove xcb shm helper
The shm helper was factored out in f4ee319c from an early experimental backend
for Wayland compositing. Neither this backend nor the shm helper are in use
anymore today and as it seems haven't been for a long time. Remove the class to
reduce the size of our code base.