refactor: move flags mechanism for enum bitfields to global namespace
Making use of our enum bitfield mechanism with bitwise operations and the
flags class outside the KWin namespace won't work when the enum class itself is
defined in a subnamespace of KWin. To circumvent this the mechanism must be
declared in the global namespace.

While this pollutes the global namespace, it is necessary to enable later usage
of our types from outside the KWin namespace, in particular when thinking of
external consumers at some point in the future.