refactor: move screen locker watcher to new desktop namespace
We introduce a new namespace called "desktop" which is supposed to contain
integration code with desktops, both common but also in the future integration
specific to different desktops. For now the screen locker watcher is moved
there, which is a common interface for desktops making mainly use of a
freedesktop.org interface.