fix: create screen edger in space child classes
The screen edger can be created late in the child class constructors. This is
necessary so on init of the edger the right override is called on the call to
the space to create an edge. For that we also store a reference to the space in
the edger and provide it as argument to the call. This also allows us to get
rid of global variable access to the space in the edger.