refactor: define invalid point in utils namespace
Move it from the utils header to a new geo header. While this references a Qt
type and the utils directory is supposed to not contain any we make a temporary
exception as later we want to define our own "point" type in this header and
then simply switch the type of the invalid point to that.