refactor: own options in Application
Make the options object non-global and own it via a smart pointer in the app.
This means we access the options object now always via the kwinApp() getter.

This is only a stopgap. In the future the options object should be split up
into subobjects and owned over the different module-platforms, composited
into a single base object.