fix: don't force QT_QPA_PLATFORM=wayland
We're setting this env variable because earlier we used it to force kwin to use
its special QPA so we need to change that back to something sensible.

However setting it to Wayland breaks apps that ship their own Qt with missing
or broken Wayland support.

Set it to be empty instead. Well-behaved Qt apps will use Wayland regardless
because of XDG_SESSION_TYPE.

BUG: 450000