feat: log Wayland window pointer address
Sometimes when debugging it is still useful to know the exact memory address of
an object. For this case don't only print out meta data of a Wayland window but
also the pointer address itself.

We replace the resource name with this, as the resource name in general does
not contain useful info for Wayland windows.