fix: destroy annexed remnants with alive leads before paint
An annexed window might be closed at any point in time independent of one of
its leaders being closed. If the non-annexed lead is still alive (or if it
was deleted without remnant) the annexed remnant should be deleted immediately
too as no remnant effects can be processed on it.

As for example wl_subsurfaces can be destroyed in arbitrary order we can't be
sure at this point in time if this is part of a teardown of the complete window
or only of a subtree of subsurfaces. We can be sure of this though at the
beginning of the next paint cycle. Therefore make destroyed annexed transient
windows by default remnants and check on the paint cycle if there is still a
non-annexed lead. Only if there is and this lead is also a remnant go on
to keep the annexed transient as it provides a texture for the remant effect,
otherwise destroy the annexed remnant immediately.