fix: destroy leaderless annexed windows directly
A wl_subsurface can be destroyed after its leader. In this case don't try to
continue the usual destroy routine because we will fail in getting the layer to
copy to the remnant.

Instead just destroy the window immediately. This makes sense as without a lead
window we can't paint any effects for the window anymore anyway.