docs: add comment on top lead check in subsurface position setter
We have to check if there is a top lead because clients are allowed to destroy
parent surfaces of subsurfaces before the subsurface itself. For one explain
this on the check.

We could however also ensure that annexed transient windows are destroyed when
the parent is. Then we wouldn't need the check because the signal disconnects
automatically. Doing so could prove difficult though as sometimes windows are
deleted while looping over all windows (e.g. in render::wayland::output). For
now add only a todo about this open question and revisit it later.