refactor: remove Screens count handling
Instead of storing the outputs count in Screens and emitting the countChanged
signal rely only on the list of outputs in base::platform.

The countChanged signal can be replaced with connections to the output added
and removed signals in base::platform. This is also more reliable than the
conglomerated countChanged signal.