refactor: store current output as pointer
Instead of using integers store the current output in the topology as a pointer
and update it when outputs are connected or disconnected.

As we now provide the central output in Toplevel and the current one in the
topology as pointers we can switch from using integers to using the output type
in functions depending on the screen. This ensures type safety and can be used
for an increased feature set in the future.