refactor: remove dmabuf buffer registry
The purpose of the list of all current dmabuf buffers in the wayland::server
class was to allow the egl_dmabuf class to be destroyed and recreated on the
fly. This was necessary for when the graphics context got reset.

But we don't do this anymore today. We can assume the same egl_dmabuf objects
keeps existing throughout the runtime and no dmabuf buffers have to be unlinked
or linked from it.