build: create wrapper link alyways in 'bin' subdirectory
The CMAKE_INSTALL_BINDIR variable can contain any path, also absolute ones.
It is assumed CMake always puts targets into a 'bin' subdirectory in the
build directory. This is what CMAKE_INSTALL_BINDIR defaults to. Replace
the variable with 'bin' to ensure we always create in and install the link
from this place.