fix(input): check for headless backend instead of device
For integration tests we need to check if we run on the headless backend on
keyboard creation. But with [1] the API function in wlroots has been removed to
do this directly on the input device. Instead now check via the wlr_backend.
This also works with the previous wlroots version.

[1] https://gitlab.freedesktop.org/wlroots/wlroots/-/commit/7d560df9