feat(effect): remove legacy X11 atom way to highlight window
Highlighting windows via the X11 atom has been deprecated since 6d7db5cb in
favor of a windowing system independent D-Bus interface. The legacy X11 way was
kept because of compatibility with KWindowSystem. That is just theoretically
though. The Plasma desktop has been ported to the D-Bus interface and is the
only relevant client for that effect. Therefore remove the legacy interface in
order to make the effect truly independent of the windowing system in use.