feat(effect): remove legacy X11 atom way to present window
Presenting windows via the X11 atom has been deprecated since f90ad31f in
favor of a windowing system independent D-Bus interface. As with the highlight
effect the legacy atom-based interface was kept but this is not necessary.
Remove the legacy interface in order to make the effect truly independent of
the windowing system in use.