refactor(effect): empty out kwineffects files
The kwineffects.cpp file is removed. We need to include the moc of the
kwinglobals header  though in some source file. We choose effects_handler.cpp
for that.

The header kwineffects.h is kept for backwards compatibility. But we move the
documentation to a README.md file in the libkwineffects subdirectory.