fix(effect): load effects in handler child class constructors
We may only call reconfigure() in the constructors of the X11 and Wayland child
class consturctors of the effects handler.

Because this loads the effects and effects may access in their constructors
various properties that require the child classes to already exist. For example
the blur effect tries to get a pointer to the Wayland display, what only the
Wayland effect handler child class can provide.