fix: render into texture framebuffer to flip for wlroots
For wlroots we need to flip the y-coordinate because internally wlroots renders
into a framebuffer. Compositors based on wlroots have to do this in general
when not using the basic render functions that wlroots provides itself. [1][2]

Before making use of the wlr_renderer we rendered to EGLSurfaces ourselves and
these already come with an inverse y-coordinate.

After integrating with the wlr_renderer and until now we compensated for the
inverse y-coordinate by flipping the image with the projection matrix. This is
the ideal solution in terms of performance and robustness but the problem is
that effects might assume a standard orientation. This is the case for the
screenshot effect which assumes it has to do the flip on its own when blitting
into a QImage.

As a suboptimal but functional solution we now don't flip via the projection
matrix anymore but always render into the separate texture framebuffer which
was before only used for enabling output rotations. Then we always do a
vertical flip when rendering from that framebuffer to the final render buffer.

This is subomtimal because writing to a texture is slower than rendering to
a render buffer.

[1] https://github.com/WayfireWM/wayfire/blob/9458f589/src/output/render-manager.cpp#L282
[2] https://git.sr.ht/~bl4ckb0ne/wxrc/tree/3eaa3b1e/item/wxrc/render/shaders.c#L53