fix: assert on restacked other window being different
It can be reasoned that these asserts are valid because in the beginning we
already checked that the "under" argument is contained in the pre-stack, so
it can't be null. And in the for-loop we only act on "other" if it belongs
to the same client as under does while window does not belong to the same
client. So by contraposition "other" can not equal "window".

This fixes a Coverity Scan warning about "under" maybe being null in the later
call to focus_chain::moveAfterClient.