feat(effect): provide blur integration
Provide a new kind of tool in the effects library reducing the footprint of
individual effects and abstracting away differences in the windowing system.

The win_integration class at the most abstract level provides hooks to
integrate in effect-specific ways with the effects system. Effects can be
added and removed via this class and effects need to provide a single update
function that the effects system can call into to update the individual effect.

Internally the effects system ensures that the effect type is integrated for
any supported windowing system.