feat(effect): provide KScreen effect integration
Analogous to the blur and other integrations introduce integration for the
KScreen effect. This effect is meant specifically for KScreen and other systems
to mask configuration changes. But at the moment it is only used by Powerdevil
which uses it to mask going into standbye. Also it is only available on X11,
not on Wayland.

We still create integration functions for it in order to reduce windowing
system specific code in individual effects. For that we reuse the generic
facilities that were introduced together with the blur integration. On Wayland
the effect is stubbed out. That is the first effect only available on one
windowing system. In general the question is if this could happen regularly and
the integration functions should return a pointer instead of a reference. For
now we assume it will only be this effect or very few others though and
continue returning references to the effect integration structs.